:white_check_mark: replace intValue with longValue in tests

+++ jspeedtest/src/test/java/fr/bmartel/speedtest/test/SpeedTestHttpsTest.java
@@ -159,8 +159,8 @@ public class SpeedTestHttpsTest extends ServerRetryTest {
         Assert.assertNotNull(mExpectedTransferRateOps);
         Assert.assertNotNull(mExpectedTransferRateBps);
 
-        Assert.assertTrue(mExpectedTransferRateBps.intValue() > 0);
-        Assert.assertTrue(mExpectedTransferRateOps.intValue() > 0);
+        Assert.assertTrue(mExpectedTransferRateBps.longValue() > 0);
+        Assert.assertTrue(mExpectedTransferRateOps.longValue() > 0);
     }
 
     /**
